fix(dynamic-branding) Fix bogus mUI dependency on mobile (#10375)
