+++ react/features/dynamic-branding/functions.web.js
@@ -1,54 +1,9 @@
 // @flow
+
 import { createMuiTheme } from '@material-ui/core/styles';
 
-import { loadConfig } from '../base/lib-jitsi-meet';
 import { font, colors, colorMap, spacing, shape, typography, breakpoints } from '../base/ui/Tokens';
 import { createColorTokens } from '../base/ui/utils';
-
-/**
- * Extracts the fqn part from a path, where fqn represents
- * tenant/roomName.
- *
- * @param {string} path - The URL path.
- * @returns {string}
- */
-export function extractFqnFromPath() {
-    const parts = window.location.pathname.split('/');
-    const len = parts.length;
-
-    return parts.length > 2 ? `${parts[len - 2]}/${parts[len - 1]}` : '';
-}
-
-/**
- * Returns the url used for fetching dynamic branding.
- *
- * @returns {string}
- */
-export async function getDynamicBrandingUrl() {
-    const config = await loadConfig(window.location.href);
-    const { dynamicBrandingUrl } = config;
-
-    if (dynamicBrandingUrl) {
-        return dynamicBrandingUrl;
-    }
-
-    const { brandingDataUrl: baseUrl } = config;
-    const fqn = extractFqnFromPath();
-
-    if (baseUrl && fqn) {
-        return `${baseUrl}?conferenceFqn=${encodeURIComponent(fqn)}`;
-    }
-}
-
-/**
- * Selector used for getting the load state of the dynamic branding data.
- *
- * @param {Object} state - Global state of the app.
- * @returns {boolean}
- */
-export function isDynamicBrandingDataLoaded(state: Object) {
-    return state['features/dynamic-branding'].customizationReady;
-}
 
 /**
  * Creates MUI branding theme based on the custom theme json.
@@ -117,13 +72,13 @@ export function createMuiBrandingTheme(customTheme: Object) {
 }
 
 /**
- * Overwrites recursively values from object 2 into object 1 based on common keys.
- * (Merges object2 into object1).
- *
- * @param {Object} obj1 - The object holding the merged values.
- * @param {Object} obj2 - The object to compare to and take values from.
- * @returns {void}
- */
+* Overwrites recursively values from object 2 into object 1 based on common keys.
+* (Merges object2 into object1).
+*
+* @param {Object} obj1 - The object holding the merged values.
+* @param {Object} obj2 - The object to compare to and take values from.
+* @returns {void}
+*/
 function overwriteRecurrsive(obj1: Object, obj2: Object) {
     Object.keys(obj2).forEach(key => {
         if (obj1.hasOwnProperty(key)) {
