+++ react/features/dynamic-branding/index.js
@@ -1,2 +1,2 @@
 export * from './actions';
-export * from './functions';
+export * from './functions.any';
